SDL3 rewrite underway 2
updated readme

+++ graphic_components/sprites.cpp
@@ -36,14 +36,14 @@ strech_bg::strech_bg(const std::string& texture, const texture_manager& tex_mgr,
 	float w, h;
 	SDL_GetTextureSize(get_tex(), &w, &h);
 	src_rect = { 0, 0, w, h };
-	screen = { 0, 0, screen_w, screen_h };
+	screen = { 0, 0, (float)screen_w, (float)screen_h };
 }
 
 void strech_bg::set_screen(int screen_w, int screen_h) {
 	float w, h;
 	SDL_GetTextureSize(get_tex(), &w, &h);
 	src_rect = { 0, 0, w, h };
-	screen = { 0, 0, screen_w, screen_h };
+	screen = { 0, 0, (float)screen_w, (float)screen_h };
 }
 
 void strech_bg::render(SDL_Renderer* ren, const Camera& cam) const {
